Minor name changes DiffCalculator refactor Minor changes to dtos
