+++ src/main/java/com/waes/diffapi/domain/calculator/DiffCalculator.java
@@ -1,0 +1,10 @@
+package com.waes.diffapi.domain.calculator;
+
+import com.waes.diffapi.domain.Diff;
+import java.util.List;
+
+public interface DiffCalculator {
+
+    List<String> calculate(Diff diff);
+
+}
